\name{configFilesDirectoryNameGet}
\alias{configFilesDirectoryNameGet}
\title{get the path of the temporary directory used for storing work files }
\description{
This directory may be an absolute or relative path and ends with a / character
}
\usage{
configFilesDirectoryNameGet()
}
\value{
Returns a string that is the path of the work directory 
}
\author{
Greg Hunt <greg@firmansyah.com>
}

\examples{
configFilesDirectoryNameGet()
}
\keyword{ ~manip }
